fix: FSD 아키텍처 공개 API 수정
