+++ src/entities/user/index.tsx
@@ -1,0 +1,1 @@
+export {};
